+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - FEATURE kumo now has the visualize cmd. Req. dot installation (#136).
 - FEATURE tenkai now has the slack notifications (#79).- FIX moved tests to pytest to improve cleanup after tests (#119).
+- FIX moved tests to pytest to improve cleanup after tests (#119).
+- FIX ramuda rollback to previous version.
 
 ## [0.0.57] - 2016-09-23
 ### Added